% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bootperturb}
\alias{bootperturb}
\title{Combined scale results resampled with uncertainty}
\description{
Results of combined calibration and main scale datasets, 
resampled 1000 times with random error introduced within bounds of 
uncertainty determined from analysis of calibration dataset
}
\examples{
data(bootperturb)
}
\author{
Estes, Treuer, Elsen, Chang, Choi, Ahmed, Caylor, Ellis
}
\keyword{data}
