% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datf}
\alias{datf}
\title{Combined scale results}
\description{
Cleaned and combined calibration and main scale datasets
}
\examples{
data(datf)
}
\author{
Estes, Treuer, Elsen, Chang, Choi, Ahmed, Caylor, Ellis
}
\keyword{data}
