% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dimbreaks}
\alias{dimbreaks}
\title{Combined scale results}
\description{
Contains multiple look-up tables for plotting and labelling 
figures in log10 spaces (aaxis1, aaxis2, taxis1, taxis2, alab1, alab2, tlab1, 
tlab2)
}
\examples{
data(datf)
}
\author{
Estes, Treuer, Elsen, Chang, Choi, Ahmed, Caylor, Ellis
}
\keyword{data}
