% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hdatstat_bt}
\alias{hdatstat_bt}
\title{Histograms of scale dimensions}
\description{
A list containing mean, 2.5, and 97.5 percentile values 
resulting from histograms applied to bootperturb for each of the four primary 
dimensions (used in making Figure 1), divided by observation type.
}
\examples{
data(hdatstat_bt)
}
\author{
Estes, Treuer, Elsen, Chang, Choi, Ahmed, Caylor, Ellis
}
\keyword{data}
