% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdat_setup.R
\name{kdat_setup}
\alias{kdat_setup}
\title{Prepares scale data for kernel density estimates}
\usage{
kdat_setup(dat, x, y, xlim, ylim, drop.oneoff = FALSE, oneoff.var = "y")
}
\arguments{
\item{dat}{Input data.table (an rbind of bootperturb)}

\item{x}{Name of variable in dat that will form x axis}

\item{xlim}{Two element vector setting limits for x axis}

\item{ylim}{Two element vector setting limits for y axis}

\item{drop.oneoff}{Should unrepeated observations be dropped (default=FALSE)}

\item{oneoff.var}{Axis containing unrepeated observations ("x" or "y")}

\item{x}{Name of variable in dat that will form y axis}
}
\value{
SpatialPointsDataFrame
}
\description{
Processes two variables from input scale data.table for KDE
}
\examples{
odt <- kdat_setup(kdat, x = "plot_res", y = "t_btwn_samp", xlim = c(-5, 8), 
                  ylim = c(-5, 5.57), drop.oneoff = TRUE)  
}
