% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdensity.R
\name{kdensity}
\alias{kdensity}
\title{Kernel density raster}
\usage{
kdensity(xmn, xmx, ymn, ymx, dxdy, pts, kwidth)
}
\arguments{
\item{xmn}{Minimum x coordinate for output grid}

\item{xmx}{Maximum x coordinate for output grid}

\item{ymn}{Minimum y coordinate for output grid}

\item{ymx}{Maximum y coordinate for output grid}

\item{dxdy}{Resolution for output grid}

\item{pts}{SpatialPointsDataFrame showing distribution of observations}

\item{kwidth}{Kernel smoothing width}
}
\value{
Raster of density values
}
\description{
Creates a raster from kernel density
}

