% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kderla}
\alias{kderla}
\title{Kernel density estimates for bivariate dimensions comparisons}
\description{
A list of kernel density rasters and various plotting 
parameters, calculated across all observations and by observational type, for
Figure 2.
}
\examples{
data(kderla)
}
\author{
Estes, Treuer, Elsen, Chang, Choi, Ahmed, Caylor, Ellis
}
\keyword{data}
