% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kderls}
\alias{kderls}
\title{Kernel density estimates for bandwidth sensitivity estimation}
\description{
A list of kernel density rasters calculated for bandwidth 
sensitivity analysis
}
\examples{
data(kderls)
}
\author{
Estes, Treuer, Elsen, Chang, Choi, Ahmed, Caylor, Ellis
}
\keyword{data}
