% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resamp_func.R
\name{resamp_func}
\alias{resamp_func}
\title{Resampling with perturbation}
\usage{
resamp_func(dat, p, iter)
}
\arguments{
\item{dat}{data.table of combined calibration and main scale results}

\item{p}{Vector of uncertainty bounds, one per assessed dimension}

\item{iter}{Number of resampling iterations}
}
\value{
List of data.table equal to length iter
}
\description{
Randomly perturbs scale values within specified uncertainty 
bounds
}
