% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{to_infin_byond}
\alias{to_infin_byond}
\title{Sets interal of non-replicated samples to 365 * 10000 days}
\usage{
to_infin_byond(tbtwn)
}
\arguments{
\item{tbtwn}{Vector of interval values}
}
\value{
Vector
}
\description{
Sets interal of non-replicated samples to 365 * 10000 days
}
